@@ -11,34 +11,33 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~13.2.1",
-    "@angular/common": "~13.2.1",
-    "@angular/compiler": "~13.2.1",
-    "@angular/core": "~13.2.1",
-    "@angular/forms": "~13.2.1",
-    "@angular/platform-browser": "~13.2.1",
-    "@angular/platform-browser-dynamic": "~13.2.1",
-    "@angular/router": "~13.2.1",
+    "@angular/animations": "^14.2.10",
+    "@angular/common": "^14.2.10",
+    "@angular/compiler": "^14.2.10",
+    "@angular/core": "^14.2.10",
+    "@angular/forms": "^14.2.10",
+    "@angular/platform-browser": "^14.2.10",
+    "@angular/platform-browser-dynamic": "^14.2.10",
+    "@angular/router": "^14.2.10",
     "@supabase/supabase-js": "^1.30.0",
     "bootstrap": "^5.1.3",
     "ngx-toastr": "^14.2.1",
     "rxjs": "~7.5.2",
     "tslib": "^2.3.1",
-
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~13.2.2",
+    "@angular-devkit/build-angular": "^14.2.9",
     "@angular-eslint/builder": "13.0.1",
     "@angular-eslint/eslint-plugin": "13.0.1",
     "@angular-eslint/eslint-plugin-template": "13.0.1",
     "@angular-eslint/schematics": "13.0.1",
     "@angular-eslint/template-parser": "13.0.1",
-    "@angular/cli": "~13.2.2",
-    "@angular/compiler-cli": "~13.2.1",
+    "@angular/cli": "^14.2.9",
+    "@angular/compiler-cli": "^14.2.10",
     "@types/jasmine": "~3.10.3",
     "@types/node": "^17.0.15",
-    "typescript": "^4.5.5",
+    "typescript": "^4.8.4",
     "@typescript-eslint/eslint-plugin": "5.10.2",
     "@typescript-eslint/parser": "5.10.2",
     "eslint": "^8.8.0",
